+++ src/color-picker/index.tsx
@@ -1,0 +1,15 @@
+export { ColorPicker } from './components/index'
+export { usePicker } from './context'
+export { useColorPicker } from './hooks/use-color-picker'
+
+export type {
+	ColorPickerProps,
+	ColorsProps,
+	Config,
+	GradientProps,
+	LocalesProps,
+	PassedConfig,
+	Styles,
+	ThemeMode,
+	ThemeProps,
+} from './types'