+++ src/index.ts
@@ -1,13 +1,48 @@
-import { ColorPicker } from './components/index'
-export { useColorPicker } from './hooks/useColorPicker'
+// Re-export all modules for convenience
+export * from './color-picker'
+export * from './input'
+export * from './input-color-picker'
+export * from './input-number-select'
+export * from './shared'
+
+// Re-export inputs with types
+export {
+	AngleInput,
+	BlurInput,
+	BorderRadiusInput,
+	BorderRadiusMultiInput,
+	FontSizeInput,
+	HeightInput,
+	LetterSpacingInput,
+	LineHeightInput,
+	OpacityInput,
+	ScaleInput,
+	SpacingInput,
+	WidthInput,
+	ZIndexInput,
+} from './inputs'
 
 export type {
-  Styles,
-  ColorsProps,
-  PassedConfig,
-  LocalesProps,
-  GradientProps,
-  ColorPickerProps,
-} from './shared/types'
+	AngleInputProps,
+	BaseInputProps,
+	BaseInputPropsWithoutUnit,
+	BlurInputProps,
+	BorderRadiusInputProps,
+	BorderRadiusMultiInputProps,
+	CommonInputProps,
+	FontSizeInputProps,
+	HeightInputProps,
+	LetterSpacingInputProps,
+	LineHeightInputProps,
+	OpacityInputProps,
+	Orientation,
+	Progression,
+	ScaleInputProps,
+	SpacingInputProps,
+	Unit,
+	WidthInputProps,
+	ZIndexInputProps,
+} from './inputs'
 
-export default ColorPicker
+// Default export for backward compatibility
+export { ColorPicker as default } from './color-picker'