+++ src/input-number-select/index.tsx
@@ -1,0 +1,2 @@
+export { InputNumberSelect } from './input-number-select'
+export type { InputNumberSelectProps, Progression } from './input-number-select'