+++ src/shared/index.ts
@@ -1,0 +1,2 @@
+export { cn } from './utils/cn'
+export { removeTrailingZeros } from './utils/remove-trailing-zeros'